# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LAZI ST-Ingestion API
  description: |
    Internal API for accessing ServiceTitan data synced to the LAZI platform.
    
    This API provides read access to master tables (synced from ServiceTitan)
    and write access via the mutation queue (changes pushed back to ServiceTitan).
    
    ## Authentication
    All endpoints require an API key passed in the `x-api-key` header.
    
    ## Rate Limiting
    - 100 requests per minute per API key
    - 1000 requests per hour per API key
    
    ## Pagination
    All list endpoints support pagination via `limit` and `offset` query parameters.
    Maximum limit is 1000 records per request.
  version: 1.0.0
  contact:
    name: LAZI Support
    email: support@lazilabs.com
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: https://st-ingestion.lazilabs.com/api
    description: Production
...
paths:
...
  /v1/pricebook/services:
    get:
      tags: [Pricebook]
      summary: List pricebook services
      description: Returns paginated list of pricebook services synced from ServiceTitan.
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/active_only'
        - $ref: '#/components/parameters/updated_after'
        - name: category_id
          in: query
          schema:
            type: integer
          description: Filter by category ID
        - name: business_unit_id
          in: query
          schema:
            type: integer
          description: Filter by business unit ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
  /v1/pricebook/services/{id}:
    get:
      tags: [Pricebook]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ServiceTitan service ID
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/pricebook/materials:
    get:
      tags: [Pricebook]
      summary: List pricebook materials
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/active_only'
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialListResponse'
  /v1/pricebook/materials/{id}:
    get:
      tags: [Pricebook]
      summary: Get material by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '404':
          description: Material not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/pricebook/equipment:
    get:
      tags: [Pricebook]
      summary: List pricebook equipment
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/active_only'
        - name: category_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentListResponse'
  /v1/pricebook/equipment/{id}:
    get:
      tags: [Pricebook]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '404':
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/pricebook/categories:
    get:
      tags: [Pricebook]
      summary: List pricebook categories
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
  /v1/pricebook/categories/{id}:
    get:
      tags: [Pricebook]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/mutations:
    get:
      tags: [Mutations]
      summary: List mutations
      description: Returns paginated list of mutations in the queue.
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed, conflict, dlq]
        - name: entity_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationListResponse'

    post:
      tags: [Mutations]
      summary: Create mutation
      description: |
        Queue a change to be pushed to ServiceTitan.
        
        The mutation will be processed by the outbound worker and sent to the
        ServiceTitan API. Use the returned mutation ID to track status.
        
        Mutations are idempotent - submitting the same payload multiple times
        will return the same mutation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMutationRequest'
            examples:
              updatePrice:
                summary: Update service price
                value:
                  entity_type: pricebook_services
                  entity_id: 12345
                  operation: update
                  payload:
                    price: 149.99
              createCustomer:
                summary: Create customer
                value:
                  entity_type: customers
                  entity_id: 0
                  operation: create
                  payload:
                    name: "John Smith"
                    type: "Residential"
                    email: "john@example.com"
      responses:
        '201':
          description: Mutation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/mutations/{id}:
    get:
      tags: [Mutations]
      summary: Get mutation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationResponse'
        '404':
          description: Mutation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
      description: Maximum records to return (max 1000)
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of records to skip
    search:
      name: search
      in: query
      schema:
        type: string
      description: Search term (searches name, code, description)
    active_only:
      name: active_only
      in: query
      schema:
        type: boolean
      description: Only return active records
    updated_after:
      name: updated_after
      in: query
      schema:
        type: string
        format: date-time
      description: Only return records synced after this timestamp
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Resource not found"
    ListMeta:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
          example: 2332
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0
        has_more:
          type: boolean
          example: true
        last_sync:
          type: string
          format: date-time
          example: "2026-01-21T18:30:00Z"
...
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 12345
        st_id:
          type: integer
          example: 12345
        display_name:
          type: string
          example: "Pool Cleaning Service"
        code:
          type: string
          example: "POOL-CLEAN-001"
        description:
          type: string
        price:
          type: number
          format: decimal
          example: 149.99
        cost:
          type: number
          format: decimal
        member_price:
          type: number
          format: decimal
        active:
          type: boolean
        category_id:
          type: integer
        business_unit_id:
          type: integer
        synced_at:
          type: string
          format: date-time
    ServiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        meta:
          $ref: '#/components/schemas/ListMeta'
    ServiceResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Service'
    Material:
      type: object
      properties:
        id:
          type: integer
        st_id:
          type: integer
        display_name:
          type: string
        code:
          type: string
        price:
          type: number
          format: decimal
        cost:
          type: number
          format: decimal
        active:
          type: boolean
        category_id:
          type: integer
        synced_at:
          type: string
          format: date-time
    MaterialListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Material'
        meta:
          $ref: '#/components/schemas/ListMeta'
    MaterialResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Material'
    Equipment:
      type: object
      properties:
        id:
          type: integer
        st_id:
          type: integer
        display_name:
          type: string
        code:
          type: string
        price:
          type: number
          format: decimal
        active:
          type: boolean
        category_id:
          type: integer
        synced_at:
          type: string
          format: date-time
    EquipmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
        meta:
          $ref: '#/components/schemas/ListMeta'
    EquipmentResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Equipment'
    Category:
      type: object
      properties:
        id:
          type: integer
        st_id:
          type: integer
        name:
          type: string
        active:
          type: boolean
        synced_at:
          type: string
          format: date-time
    CategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        meta:
          $ref: '#/components/schemas/ListMeta'
    CategoryResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Category'
...
    Mutation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
          example: "pricebook_services"
        entity_id:
          type: string
          example: "12345"
        operation:
          type: string
          enum: [create, update, delete]
        status:
          type: string
          enum: [pending, processing, completed, failed, conflict, dlq]
        payload:
          type: object
        error_message:
          type: string
        retry_count:
          type: integer
        idempotency_key:
          type: string
        initiated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    MutationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Mutation'
        meta:
          $ref: '#/components/schemas/ListMeta'
    MutationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Mutation'
        message:
          type: string
    CreateMutationRequest:
      type: object
      required: [entity_type, entity_id, operation, payload]
      properties:
        entity_type:
          type: string
          description: Type of entity (e.g., pricebook_services, customers)
          example: "pricebook_services"
        entity_id:
          type: integer
          description: ServiceTitan entity ID (0 for create operations)
          example: 12345
        operation:
          type: string
          enum: [create, update, delete]
          description: Operation to perform
        payload:
          type: object
          description: Data to send to ServiceTitan
          example:
            price: 149.99
        initiated_by:
          type: string
          description: User or system identifier
          example: "user-123"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
security:
  - ApiKeyAuth: []